API

API

Types and Constants

VTKView.DataSet — Method.

DataSet(this)

Create a data set.

source

VTKView.Frame — Method.

Frame(this)

Create Frame (corresponds to desktop window).

source

VTKView.GridView — Method.

GridView(this)

Create a view for showing 2D or 3D grids.

source

VTKView.ScalarView — Method.

ScalarView(this)

Create a view for showing 2D or 3D scalar data.

source

VTKView.VectorView — Method.

VectorView(this)

Create a view for showing 2D or 3D vector data.

source

VTKView.XYPlot — Method.

XYPlot(this)

Create a view for showing X-Y plots.

source

Methods

Base.Multimedia.display — Method.

display(this)

Display frame content and continue.

source

VTKView.addplot! — Method.

addplot!(this, X, Y)

Add one paticular plot ("series") to XYPlot object.

source

VTKView.addpointscalar! — Method.

addpointscalar!(this, X, name)

Add piece wise linear function with identifier name to dataset.

source

VTKView.addview! — Method.

addview!(this, fig, pos)

Add view to frame.

source

VTKView.addview! — Method.

addview!(this, fig)

Add view to frame.

source

VTKView.addview! — Method.

addview!(this, fig, pos)

Add view to frame.

source

VTKView.addview! — Method.

addview!(this, fig)

Add view to frame.

source

VTKView.addview! — Method.

addview!(this, fig, pos)

Add view to frame.

source

VTKView.addview! — Method.

addview!(this, fig)

Add view to frame.

source

VTKView.addview! — Method.

addview!(this, fig, pos)

Add view to frame.

source

VTKView.addview! — Method.

addview!(this, fig)

Add view to frame.

source

VTKView.adjust_labels! — Method.

adjust_labels!(this, b)

Toggle automatic label adjustment.

source

VTKView.axescolor! — Method.

axescolor!(this, r, g, b)

Color of axes.

source

VTKView.clear! — Method.

clear!(this)

Clear frame content.

source

VTKView.clear! — Method.

clear!(this)

Clear all data from plot.

source

VTKView.data! — Method.

data!(this, dataset)

Set data to be shown to grid defined in dataset.

source

VTKView.data! — Method.

data!(this, dataset, name)

Set data to be shown to point scalar name in dataset.

source

VTKView.frametitle! — Method.

frametitle!(this, title)

Remove view from frame.

source

VTKView.gridcolor! — Method.

gridcolor!(this, r, g, b)

Color of background grid.

source

VTKView.interact — Method.

interact(this)

Display frame content and enter interaction mode.

source

VTKView.isolevels! — Method.

isolevels!(this, V)

Set isolevels for isocontour and isosurface plot.

source

VTKView.layout! — Method.

layout!(this, x, y)

Remove view from frame.

source

VTKView.legendposition! — Method.

legendposition!(this, x, y)

Set legend position.

source

VTKView.legendsize! — Method.

legendsize!(this, w, h)

Sed legend size.

source

VTKView.linewidth! — Method.

linewidth!(this, w)

Linewidth of next plot.

source

VTKView.markersize! — Method.

markersize!(this, s)

Marker size of next plot.

source

VTKView.numberofxlabels! — Method.

numberofxlabels!(this, n)

Number of labels on X axis.

source

VTKView.numberofylabels! — Method.

numberofylabels!(this, n)

Number of labels on Y axis.

source

VTKView.plotcolor! — Method.

plotcolor!(this, r, g, b)

Color of next plot.

source

VTKView.plotlegend! — Method.

plotlegend!(this, legend)

Legend of next plot.

source

VTKView.plotlinetype! — Method.

plotlinetype!(this, ltype)

Line type of next plot.

source

VTKView.plotmarkertype! — Method.

plotmarkertype!(this, mtype)

Marker type of next plot.

source

VTKView.position! — Method.

position!(this, x, y)

Remove view from frame.

source

VTKView.rectilineargrid! — Method.

rectilineargrid!(this, X, Y, Z)

Create 3D rectilinear grid in dataset.

source

VTKView.rectilineargrid! — Method.

rectilineargrid!(this, X, Y)

Create 2D rectilinear grid in dataset.

source

VTKView.remove! — Method.

remove!(this, fig)

Remove view from frame.

source

VTKView.remove! — Method.

remove!(this, fig)

Remove view from frame.

source

VTKView.remove! — Method.

remove!(this, fig)

Remove view from frame.

source

VTKView.remove! — Method.

remove!(this, fig)

Remove view from frame.

source

VTKView.show_grid! — Method.

show_grid!(this, b)

Toggle display of background grid in plot.

source

VTKView.show_isosurfaces! — Method.

show_isosurfaces!(this, b)

Toggle display of isosurfaces in 3D plot.

source

VTKView.show_legend! — Method.

show_legend!(this, b)

Toggle display of legend in plot.

source

VTKView.simplexgrid! — Method.

simplexgrid!(this, Points, Cells)

Create simplex grid in dataset.

source

VTKView.size! — Method.

size!(this, x, y)

Remove view from frame.

source

VTKView.startvideo — Method.

startvideo(this, fname)

Remove view from frame.

source

VTKView.stopvideo — Method.

stopvideo(this)

Remove view from frame.

source

VTKView.title! — Method.

title!(this, title)

Set plot title.

source

VTKView.windowtitle! — Method.

windowtitle!(this, title)

Remove view from frame.

source

VTKView.writepng — Method.

writepng(this, fname)

Remove view from frame.

source

VTKView.writevtk — Method.

writevtk(this, fname; ftype)

Write data set content to vtk file.

source

VTKView.xrange! — Method.

xrange!(this, x0, x1)

Range of X coordinates.

source

VTKView.xtitle! — Method.

xtitle!(this, title)

X Axis title.

source

VTKView.yrange! — Method.

yrange!(this, y0, y1)

Range of Y coordinates.

source

VTKView.ytitle! — Method.

ytitle!(this, title)

Y Axis title.

source

VTKView.StaticFrame — Method.

StaticFrame()

Return static frame, and create it if it doesnt exist

source

VTKView.delaunay — Method.

delaunay(inpoints)

Create Delaunay triangulation from set of points

source

VTKView.destroy_static_frame — Method.

destroy_static_frame()

Destroy static frame

source