API

API

Types and Constants

VTKView.DataSet — Method.

DataSet(this::VTKView.DataSet) -> VTKView.DataSet

Create a data set.

source

VTKView.Frame — Method.

Frame(this::VTKView.Frame) -> VTKView.Frame

Create Frame (corresponds to desktop window).

source

VTKView.GridView — Method.

GridView(this::VTKView.GridView) -> VTKView.GridView

Create a view for showing 2D or 3D grids.

source

VTKView.ScalarView — Method.

ScalarView(this::VTKView.ScalarView) -> VTKView.ScalarView

Create a view for showing 2D or 3D scalar data.

source

VTKView.VectorView — Method.

VectorView(this::VTKView.VectorView) -> VTKView.VectorView

Create a view for showing 2D or 3D vector data.

source

VTKView.XYPlot — Method.

XYPlot(this::VTKView.XYPlot) -> VTKView.XYPlot

Create a view for showing X-Y plots.

source

Methods

Base.Multimedia.display — Method.

display(frame::VTKView.Frame)

Display frame content and continue.

source

VTKView.addplot! — Method.

addplot!(this::VTKView.XYPlot, X::Array{Float64,1}, Y::Array{Float64,1})

Add one paticular plot ("series") to XYPlot object.

source

VTKView.addview! — Method.

addview!(frame::VTKView.Frame, gridview::VTKView.GridView, pos::Int64)

Add view to frame.

source

VTKView.addview! — Method.

addview!(frame::VTKView.Frame, gridview::VTKView.GridView)

Add view to frame.

source

VTKView.addview! — Method.

addview!(frame::VTKView.Frame, scalarview::VTKView.ScalarView, pos::Int64)

Add view to frame.

source

VTKView.addview! — Method.

addview!(frame::VTKView.Frame, scalarview::VTKView.ScalarView)

Add view to frame.

source

VTKView.addview! — Method.

addview!(frame::VTKView.Frame, vectorview::VTKView.VectorView, pos::Int64)

Add view to frame.

source

VTKView.addview! — Method.

addview!(frame::VTKView.Frame, vectorview::VTKView.VectorView)

Add view to frame.

source

VTKView.addview! — Method.

addview!(frame::VTKView.Frame, xyplot::VTKView.XYPlot, pos::Int64)

Add view to frame.

source

VTKView.addview! — Method.

addview!(frame::VTKView.Frame, xyplot::VTKView.XYPlot)

Add view to frame.

source

VTKView.adjust_labels! — Method.

adjust_labels!(this::VTKView.XYPlot, b::Bool)

Toggle automatic label adjustment.

source

VTKView.axescolor! — Method.

axescolor!(this::VTKView.XYPlot, r::Any, g::Any, b::Any)

Color of axes.

source

VTKView.clear! — Method.

clear!(frame::VTKView.Frame)

Clear frame content.

source

VTKView.clear! — Method.

clear!(this::VTKView.XYPlot)

Clear all data from plot.

source

VTKView.data! — Method.

data!(this::VTKView.GridView, dataset::VTKView.DataSet)

Set data to be shown to grid defined in dataset.

source

VTKView.data! — Method.

data!(this::VTKView.ScalarView, dataset::VTKView.DataSet, name::Any)

Set data to be shown to point scalar name in dataset.

source

VTKView.data! — Method.

data!(this::VTKView.VectorView, dataset::VTKView.DataSet, name::Any)

Set data to be shown to point scalar name in dataset.

source

VTKView.frametitle! — Method.

frametitle!(frame::VTKView.Frame, title::Any)

Set title in frame

source

VTKView.gridcolor! — Method.

gridcolor!(this::VTKView.XYPlot, r::Any, g::Any, b::Any)

Color of background grid.

source

VTKView.interact — Method.

interact(frame::VTKView.Frame)

Display frame content and enter interaction mode.

source

VTKView.isolevels! — Method.

isolevels!(this::VTKView.ScalarView, V::Array{Float64,1})

Set isolevels for isocontour and isosurface plot.

source

VTKView.layout! — Method.

layout!(frame::VTKView.Frame, nrows::Any, ncol::Any)

Set frame layout.

source

VTKView.legendposition! — Method.

legendposition!(this::VTKView.XYPlot, x::Any, y::Any)

Set legend position.

source

VTKView.legendsize! — Method.

legendsize!(this::VTKView.XYPlot, w::Any, h::Any)

Sed legend size.

source

VTKView.linewidth! — Method.

linewidth!(this::VTKView.XYPlot, w::Any)

Linewidth of next plot.

source

VTKView.markersize! — Method.

markersize!(this::VTKView.XYPlot, s::Any)

Marker size of next plot.

source

VTKView.numberofxlabels! — Method.

numberofxlabels!(this::VTKView.XYPlot, n::Any)

Number of labels on X axis.

source

VTKView.numberofylabels! — Method.

numberofylabels!(this::VTKView.XYPlot, n::Any)

Number of labels on Y axis.

source

VTKView.plotcolor! — Method.

plotcolor!(this::VTKView.XYPlot, r::Any, g::Any, b::Any)

Color of next plot.

source

VTKView.plotlegend! — Method.

plotlegend!(this::VTKView.XYPlot, legend::Any)

Legend of next plot.

source

VTKView.plotlinetype! — Method.

plotlinetype!(this::VTKView.XYPlot, ltype::Any)

Line type of next plot.

source

VTKView.plotmarkertype! — Method.

plotmarkertype!(this::VTKView.XYPlot, mtype::Any)

Marker type of next plot.

source

VTKView.pointscalar! — Method.

pointscalar!(this::VTKView.DataSet, X::Array{Float64,1}, name::Any)

Add piece wise linear function with identifier name to dataset.

source

VTKView.pointvector! — Method.

pointvector!(this::VTKView.DataSet, UVW::Array{Float64,2}, name::Any)

Add piece wise linear function with identifier name to dataset.

source

VTKView.position! — Method.

position!(frame::VTKView.Frame, x::Any, y::Any)

Set frame position on screen.

source

VTKView.rectilineargrid! — Method.

rectilineargrid!(this::VTKView.DataSet, X::Array{Float64,1}, Y::Array{Float64,1}, Z::Array{Float64,1})

Create 3D rectilinear grid in dataset.

source

VTKView.rectilineargrid! — Method.

rectilineargrid!(this::VTKView.DataSet, X::Array{Float64,1}, Y::Array{Float64,1})

Create 2D rectilinear grid in dataset.

source

VTKView.remove! — Method.

remove!(frame::VTKView.Frame, gridview::VTKView.GridView)

Remove view from frame.

source

VTKView.remove! — Method.

remove!(frame::VTKView.Frame, scalarview::VTKView.ScalarView)

Remove view from frame.

source

VTKView.remove! — Method.

remove!(frame::VTKView.Frame, vectorview::VTKView.VectorView)

Remove view from frame.

source

VTKView.remove! — Method.

remove!(frame::VTKView.Frame, xyplot::VTKView.XYPlot)

Remove view from frame.

source

VTKView.show_grid! — Method.

show_grid!(this::VTKView.XYPlot, b::Bool)

Toggle display of background grid in plot.

source

VTKView.show_isosurfaces! — Method.

show_isosurfaces!(this::VTKView.ScalarView, b::Bool)

Toggle display of isosurfaces in 3D plot.

source

VTKView.show_legend! — Method.

show_legend!(this::VTKView.XYPlot, b::Bool)

Toggle display of legend in plot.

source

VTKView.simplexgrid! — Method.

simplexgrid!(this::VTKView.DataSet, Points::Array{Float64,2}, Cells::Array{Int32,2})

Create simplex grid in dataset.

source

VTKView.size! — Method.

size!(frame::VTKView.Frame, x::Any, y::Any)

Set frame size.

source

VTKView.startvideo — Method.

startvideo(frame::VTKView.Frame, fname::Any)

Start video recording.

source

VTKView.stopvideo — Method.

stopvideo(frame::VTKView.Frame)

Stop video recording

source

VTKView.title! — Method.

title!(this::VTKView.XYPlot, title::Any)

Set plot title.

source

VTKView.windowtitle! — Method.

windowtitle!(frame::VTKView.Frame, title::Any)

Set title of window corresponding to frame

source

VTKView.writepng — Method.

writepng(frame::VTKView.Frame, fname::Any)

Write frame content to png file.

source

VTKView.writevtk — Method.

writevtk(this::VTKView.DataSet, fname::Any; ftype)

Write data set content to vtk file.

source

VTKView.xrange! — Method.

xrange!(this::VTKView.XYPlot, x0::Any, x1::Any)

Range of X coordinates.

source

VTKView.xtitle! — Method.

xtitle!(this::VTKView.XYPlot, title::Any)

X Axis title.

source

VTKView.yrange! — Method.

yrange!(this::VTKView.XYPlot, y0::Any, y1::Any)

Range of Y coordinates.

source

VTKView.ytitle! — Method.

ytitle!(this::VTKView.XYPlot, title::Any)

Y Axis title.

source

VTKView.StaticFrame — Method.

StaticFrame()

Return static frame, and create it if it doesnt exist

source

VTKView.delaunay — Method.

delaunay(inpoints)

Create Delaunay triangulation from set of points

source

VTKView.destroy_static_frame — Method.

destroy_static_frame()

Destroy static frame

source