API

Types and Constants

VTKView.DataSet — Method

DataSet(this)

Create a data set.

VTKView.Frame — Method

Frame(this)

Create Frame (corresponds to desktop window).

VTKView.GridView — Method

GridView(this)

Create a view for showing 2D or 3D grids.

VTKView.ScalarView — Method

ScalarView(this)

Create a view for showing 2D or 3D scalar data.

VTKView.VectorView — Method

VectorView(this)

Create a view for showing 2D or 3D vector data.

VTKView.XYPlot — Method

XYPlot(this)

Create a view for showing X-Y plots.

Methods

Base.Multimedia.display — Method

display(this)

Display frame content and continue.

VTKView.addplot! — Method

addplot!(this, X, Y)

Add one paticular plot ("series") to XYPlot object.

VTKView.addpointscalar! — Method

addpointscalar!(this, X, name)

Add piece wise linear function with identifier name to dataset.

VTKView.addview! — Method

addview!(this, fig, pos)

Add view to frame.

VTKView.addview! — Method

addview!(this, fig)

Add view to frame.

VTKView.addview! — Method

addview!(this, fig, pos)

Add view to frame.

VTKView.addview! — Method

addview!(this, fig)

Add view to frame.

VTKView.addview! — Method

addview!(this, fig, pos)

Add view to frame.

VTKView.addview! — Method

addview!(this, fig)

Add view to frame.

VTKView.addview! — Method

addview!(this, fig, pos)

Add view to frame.

VTKView.addview! — Method

addview!(this, fig)

Add view to frame.

VTKView.adjust_labels! — Method

adjust_labels!(this, b)

Toggle automatic label adjustment.

VTKView.axescolor! — Method

axescolor!(this, r, g, b)

Color of axes.

VTKView.clear! — Method

clear!(this)

Clear frame content.

VTKView.clear! — Method

clear!(this)

Clear all data from plot.

VTKView.data! — Method

data!(this, dataset)

Set data to be shown to grid defined in dataset.

VTKView.data! — Method

data!(this, dataset, name)

Set data to be shown to point scalar name in dataset.

VTKView.frametitle! — Method

frametitle!(this, title)

Remove view from frame.

VTKView.gridcolor! — Method

gridcolor!(this, r, g, b)

Color of background grid.

VTKView.interact — Method

interact(this)

Display frame content and enter interaction mode.

VTKView.isolevels! — Method

isolevels!(this, V)

Set isolevels for isocontour and isosurface plot.

VTKView.layout! — Method

layout!(this, x, y)

Remove view from frame.

VTKView.legendposition! — Method

legendposition!(this, x, y)

Set legend position.

VTKView.legendsize! — Method

legendsize!(this, w, h)

Sed legend size.

VTKView.linewidth! — Method

linewidth!(this, w)

Linewidth of next plot.

VTKView.markersize! — Method

markersize!(this, s)

Marker size of next plot.

VTKView.numberofxlabels! — Method

numberofxlabels!(this, n)

Number of labels on X axis.

VTKView.numberofylabels! — Method

numberofylabels!(this, n)

Number of labels on Y axis.

VTKView.plotcolor! — Method

plotcolor!(this, r, g, b)

Color of next plot.

VTKView.plotlegend! — Method

plotlegend!(this, legend)

Legend of next plot.

VTKView.plotlinetype! — Method

plotlinetype!(this, ltype)

Line type of next plot.

VTKView.plotmarkertype! — Method

plotmarkertype!(this, mtype)

Marker type of next plot.

VTKView.position! — Method

position!(this, x, y)

Remove view from frame.

VTKView.rectilineargrid! — Method

rectilineargrid!(this, X, Y, Z)

Create 3D rectilinear grid in dataset.

VTKView.rectilineargrid! — Method

rectilineargrid!(this, X, Y)

Create 2D rectilinear grid in dataset.

VTKView.remove! — Method

remove!(this, fig)

Remove view from frame.

VTKView.remove! — Method

remove!(this, fig)

Remove view from frame.

VTKView.remove! — Method

remove!(this, fig)

Remove view from frame.

VTKView.remove! — Method

remove!(this, fig)

Remove view from frame.

VTKView.show_grid! — Method

show_grid!(this, b)

Toggle display of background grid in plot.

VTKView.show_isosurfaces! — Method

show_isosurfaces!(this, b)

Toggle display of isosurfaces in 3D plot.

VTKView.show_legend! — Method

show_legend!(this, b)

Toggle display of legend in plot.

VTKView.simplexgrid! — Method

simplexgrid!(this, Points, Cells)

Create simplex grid in dataset.

VTKView.size! — Method

size!(this, x, y)

Remove view from frame.

VTKView.startvideo — Method

startvideo(this, fname)

Remove view from frame.

VTKView.stopvideo — Method

stopvideo(this)

Remove view from frame.

VTKView.title! — Method

title!(this, title)

Set plot title.

VTKView.windowtitle! — Method

windowtitle!(this, title)

Remove view from frame.

VTKView.writepng — Method

writepng(this, fname)

Remove view from frame.

VTKView.writevtk — Method

writevtk(this, fname; ftype)

Write data set content to vtk file.

VTKView.xrange! — Method

xrange!(this, x0, x1)

Range of X coordinates.

VTKView.xtitle! — Method

xtitle!(this, title)

X Axis title.

VTKView.yrange! — Method

yrange!(this, y0, y1)

Range of Y coordinates.

VTKView.ytitle! — Method

ytitle!(this, title)

Y Axis title.

VTKView.StaticFrame — Method

StaticFrame()

Return static frame, and create it if it doesnt exist

VTKView.delaunay — Method

delaunay(inpoints)

Create Delaunay triangulation from set of points

VTKView.destroy_static_frame — Method

destroy_static_frame()

Destroy static frame